Title: 사각형 채우기 3, Time: 143ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240922/사각형 채우기 3/rectangle-fill-3.java
@@ -7,7 +7,7 @@ public class Main {
 		Scanner sc = new Scanner(System.in);
 		int n = sc.nextInt();
 		
-		int[] dp = new int[1001];
+		long[] dp = new long[1001];
 		
 		dp[0] = 1;
 		dp[1] = 2;
